修正raw data长度不正确的问题

+++ App/Driver/datastruct.h
@@ -83,8 +83,8 @@ typedef struct{
     uint8_t name_index;
     Sens_data_struct data1; //这俩描述了要如何处理raw数据
     Sens_data_struct data2; //要是C支持柯里化函数，这里直接用委托就行了
-    uint16_t  data1_raw;
-    uint16_t  data2_raw;
+    uint32_t  data1_raw;
+    uint32_t  data2_raw;
 }Sens_dev_desc;// 数据链路级传感器描述符
 
 typedef struct{
